==> notebooks/06-summarize-results.ipynb <==
import pandas as pd

nb = pd.read_csv("../results/NB-idf.csv")
svm = pd.read_csv("../results/svm-idf.csv")
ree = pd.read_csv("../results/ensemble-REE.csv")
roberta = pd.read_csv("../results/roberta.csv")
cnn = pd.read_csv("../results/cnn.csv")

## NB

nb.groupby(nb.category).mean().round(3)

## SVM

svm.groupby(svm.category).mean().round(3)

## Roberta

roberta.groupby(roberta.category).mean().round(3)

## CNN

cnn.groupby(cnn.category).mean().round(3)

## REE

ree.groupby(ree.category).mean().round(3)

nb['model'] = 'Naive Bayes'
svm['model'] = 'SVM'
ree['model'] = 'REE Ensemble'
roberta['model'] = 'RoBERTa'
cnn['model'] = 'CNN'

combined = pd.concat([nb,svm,ree,roberta,cnn])

combined.groupby(combined.model).mean().round(3)
